adding /auth route
