+++ src/App.js
@@ -1,8 +1,13 @@
 import './App.css';
 import { ThemeProvider } from '@mui/material/styles';
 import themeOptions from './theme/themeOptions';
-import Login from './views/home/components/login/login';
-import Home from './views/home/home';
+import Auth from './views/auth/auth';
+import {
+  BrowserRouter as Router,
+  Switch,
+  Route,
+  Link
+} from "react-router-dom";
 
 
 function App() {
@@ -10,7 +15,18 @@ function App() {
   return (
     <ThemeProvider theme={themeOptions}>
       <div className="App">
-        <Home />
+        <Router>
+          <div>
+            <Switch>
+              <Route path="/auth">
+                <Auth />
+              </Route>
+              <Route path="/">
+                <p>This is home</p>
+              </Route>
+            </Switch>
+          </div>
+        </Router>
       </div>
     </ThemeProvider>
   );
