+++ src/views/auth/components/signup/signup.jsx
@@ -1,17 +1,14 @@
-import React, { createRef, useRef, useState } from 'react'
+import React from 'react'
 import { motion } from 'framer-motion'
-import { Typography, Button, Grid, Box } from '@mui/material'
+import { Button } from '@mui/material'
 import { Input } from '../../../../theme/components'
-import { certiWhite, certiYellow } from '../../../../assets/images'
-import { FingerPrint, FingerPrint2 } from '../../../../assets/lottie'
 
-
-export default function Login() {
+export default function Signup() {
     return (
         <motion.div
-            animate={{ rotateY: 0 }}
-            transition={{ duration: 0.5 }}
-            initial={{ rotateY: -90 }}
+            animate={{ opacity: 1 }}
+            transition={{ duration: 0.2 }}
+            initial={{ opacity: 0 }}
             style={{
                 display: "flex",
                 alignItems: "center",
@@ -22,9 +19,10 @@ export default function Login() {
             }}
         >
 
+            <Input label="Name" required={true} />
             <Input label="Email" required={true} />
             <Input label="Password" type="password" required={true} />
-            <Button size="large" color="primary" variant="contained">Sign In</Button>
+            <Button size="large" color="primary" variant="contained">Sign Up</Button>
         </motion.div>
     )
 }
